Add new standards
The new guidelines for commenting style and
folder structure were applied.
