+++ .gitignore
@@ -90,3 +90,6 @@ typings/
 
 *~
 .DS_Store
+
+# SPDX-License-Identifier: (EUPL-1.2)
+# Copyright © 2019 Werbeagentur Christian Aichner
