+++ src/components/pages/index.js
@@ -1,0 +1,12 @@
+//> Components
+// Import all components to export them for easy access from parent components
+import IntroPage from './Intro';
+
+export {
+    IntroPage,
+}
+
+/** 
+ * SPDX-License-Identifier: (EUPL-1.2)
+ * Copyright © 2019 Werbeagentur Christian Aichner
+ */
