+++ src/index.scss
@@ -1,0 +1,13 @@
+//> Utilities
+// Variables
+@import 'utilities/variables';
+// Global CSS
+@import 'utilities/base';
+
+//> Components
+// Components are rendered internally to improve performance
+
+/** 
+ * SPDX-License-Identifier: (EUPL-1.2)
+ * Copyright © 2019 Werbeagentur Christian Aichner
+ */
