+++ src/registerServiceWorker.js
@@ -106,3 +106,8 @@ export function unregister() {
     });
   }
 }
+
+/** 
+ * SPDX-License-Identifier: (EUPL-1.2)
+ * Copyright © 2019 Werbeagentur Christian Aichner
+ */
