--- src/style.scss
@@ -1,5 +1,0 @@
-/* Global css for displaying purposes */
-
-html, body{
-    background: black;
-}
