+++ src/utilities/_base.scss
@@ -1,0 +1,8 @@
+html, body{
+    background: black;
+}
+
+/** 
+ * SPDX-License-Identifier: (EUPL-1.2)
+ * Copyright © 2019 Werbeagentur Christian Aichner
+ */
