+++ src/utilities/_variables.scss
@@ -1,3 +1,4 @@
+// Standard colors
 $primary: #cce5ff;
 $secondary: #e2e3e5;
 $success: #d4edda;
@@ -6,9 +7,15 @@ $warning: #fff3cd;
 $info: #d1ecf1;
 $light: #fefefe;
 $dark: #d6d8d9;
-// New
+
+// Outrun colors
 $hot: #FF1690;
 $space: #36CDC4;
 $sun: #F4D676;
 $deep: #5C2686;
 $crazy: linear-gradient(hotpink, tomato);
+
+/** 
+ * SPDX-License-Identifier: (EUPL-1.2)
+ * Copyright © 2019 Werbeagentur Christian Aichner
+ */
